fix: get rid of circular dep
