+++ apps/api/src/sandbox/services/runner.service.ts
@@ -3,10 +3,10 @@
  * SPDX-License-Identifier: AGPL-3.0
  */
 
-import { Injectable, Logger, NotFoundException, forwardRef, Inject } from '@nestjs/common'
+import { Injectable, Logger, NotFoundException } from '@nestjs/common'
 import { InjectRepository } from '@nestjs/typeorm'
 import { Cron, CronExpression } from '@nestjs/schedule'
-import { FindOptionsWhere, In, IsNull, MoreThanOrEqual, Not, Repository } from 'typeorm'
+import { FindOptionsWhere, In, MoreThanOrEqual, Not, Repository } from 'typeorm'
 import { Runner } from '../entities/runner.entity'
 import { CreateRunnerDto } from '../dto/create-runner.dto'
 import { SandboxClass } from '../enums/sandbox-class.enum'
@@ -28,8 +28,6 @@ import { WithInstrumentation } from '../../common/decorators/otel.decorator'
 import { InjectRedis } from '@nestjs-modules/ioredis'
 import { Redis } from 'ioredis'
 import { SandboxDesiredState } from '../enums/sandbox-desired-state.enum'
-import { BackupState } from '../enums/backup-state.enum'
-import { SnapshotService } from './snapshot.service'
 
 @Injectable()
 export class RunnerService {
@@ -47,8 +45,6 @@ export class RunnerService {
     private readonly configService: TypedConfigService,
     @InjectRedis()
     private readonly redis: Redis,
-    @Inject(forwardRef(() => SnapshotService))
-    private readonly snapshotService: SnapshotService,
   ) {}
 
   async create(createRunnerDto: CreateRunnerDto): Promise<Runner> {
@@ -193,7 +189,7 @@ export class RunnerService {
       return
     }
 
-    // Don't change state if runner is decommissioned or draining
+    // Don't change state if runner is decommissioned
     if (runner.state === RunnerState.DECOMMISSIONED) {
       this.logger.debug(`Runner ${runnerId} is decommissioned, not updating state`)
       return
@@ -362,100 +358,6 @@ export class RunnerService {
             }
           } catch (e) {
             this.logger.error(`Error checking draining runner ${runner.id}`, e)
-          }
-        }),
-      )
-    } finally {
-      await this.redisLockProvider.unlock(lockKey)
-    }
-  }
-
-  // Moves stopped sandboxes' backups to another runner to prepare reassignment
-  @Cron(CronExpression.EVERY_10_SECONDS, { name: 'migrate-draining-runner-snapshots', waitForCompletion: true })
-  @LogExecution('migrate-draining-runner-snapshots')
-  @WithInstrumentation()
-  private async handleMigrateDrainingRunnerSnapshots() {
-    const lockKey = 'migrate-draining-runner-snapshots'
-    const hasLock = await this.redisLockProvider.lock(lockKey, 60)
-    if (!hasLock) {
-      return
-    }
-
-    try {
-      const drainingRunners = await this.runnerRepository.find({
-        where: {
-          isDraining: true,
-        },
-      })
-
-      this.logger.debug(`Checking ${drainingRunners.length} draining runners for snapshot migration`)
-
-      await Promise.allSettled(
-        drainingRunners.map(async (runner) => {
-          try {
-            const sandboxes = await this.sandboxRepository.find({
-              where: {
-                runnerId: runner.id,
-                state: SandboxState.STOPPED,
-                desiredState: SandboxDesiredState.STOPPED,
-                backupState: BackupState.COMPLETED,
-                backupSnapshot: Not(IsNull()),
-              },
-              take: 100,
-            })
-
-            this.logger.debug(
-              `Found ${sandboxes.length} eligible sandboxes on draining runner ${runner.id} for snapshot migration`,
-            )
-
-            await Promise.allSettled(
-              sandboxes.map(async (sandbox) => {
-                const sandboxLockKey = `draining-runner-snapshot-migration:${sandbox.id}`
-                const hasSandboxLock = await this.redisLockProvider.lock(sandboxLockKey, 3600)
-                if (!hasSandboxLock) {
-                  return
-                }
-
-                try {
-                  // Get an available runner in the same region with the same class
-                  const targetRunner = await this.getRandomAvailableRunner({
-                    regions: [sandbox.region],
-                    sandboxClass: sandbox.class,
-                    excludedRunnerIds: [runner.id],
-                  })
-
-                  // Check if snapshot runner entry already exists
-                  const existingEntry = await this.getSnapshotRunner(targetRunner.id, sandbox.backupSnapshot)
-                  if (existingEntry) {
-                    this.logger.debug(
-                      `Snapshot runner entry already exists for runner ${targetRunner.id} and snapshot ${sandbox.backupSnapshot}`,
-                    )
-                    // Do not unlock to avoid duplicates
-                    return
-                  }
-
-                  // Create snapshot runner entry on the target runner
-                  await this.createSnapshotRunnerEntry(targetRunner.id, sandbox.backupSnapshot)
-                  this.snapshotService.pullSnapshotToRunner(sandbox.backupSnapshot, targetRunner)
-
-                  this.logger.log(
-                    `Created snapshot runner entry for sandbox ${sandbox.id} backup ${sandbox.backupSnapshot} on runner ${targetRunner.id} (migrating from draining runner ${runner.id})`,
-                  )
-                  await this.redisLockProvider.unlock(sandboxLockKey)
-                } catch (e) {
-                  if (e instanceof BadRequestError && e.message === 'No available runners') {
-                    this.logger.warn(
-                      `No available runners found in region ${sandbox.region} for sandbox ${sandbox.id} snapshot migration`,
-                    )
-                  } else {
-                    this.logger.error(`Error migrating snapshot for sandbox ${sandbox.id}`, e)
-                  }
-                  await this.redisLockProvider.unlock(sandboxLockKey)
-                }
-              }),
-            )
-          } catch (e) {
-            this.logger.error(`Error processing draining runner ${runner.id} for snapshot migration`, e)
           }
         }),
       )
